refactor(credit): modularize pipeline_engine + pipeline_intelligence → pipeline/ package (Wave 1 PR #5)
Split pipeline_engine.py (1030 LOC) + pipeline_intelligence.py (385 LOC)
into edgar-style package with 7 domain modules:
- models.py: constants, status enums, DD document config (leaf)
- screening.py: _retrieve_deal_context, _compute_missing_documents, compute_completeness_score
- validation.py: _validate_output, _validate_memo
- persistence.py: _set_intelligence_status, _write_research_output, _write_derived_fields
- discovery.py: discover_pipeline_deals, aggregate_deal_documents
- monitoring.py: run_pipeline_monitoring, deprecated stubs
- intelligence.py: generate_pipeline_intelligence, generate_all_pending
- service.py: run_pipeline_ingest (orchestrator)

Standard imports (no PEP 562). structlog replaces stdlib logging.
All callers updated: ai_engine/__init__.py, pipeline_ingest_runner.py,
domain_ai_engine.py, extraction.py.
