+++ backend/tests/test_vertical_engines.py
@@ -44,8 +44,7 @@ class TestCreditModuleStructure:
         "vertical_engines.credit.sponsor",
         "vertical_engines.credit.memo_chapter_engine",
         "vertical_engines.credit.memo_book_generator",
-        "vertical_engines.credit.pipeline_engine",
-        "vertical_engines.credit.pipeline_intelligence",
+        "vertical_engines.credit.pipeline",
         "vertical_engines.credit.portfolio_intelligence",
         "vertical_engines.credit.retrieval",
         "vertical_engines.credit.tone_normalizer",
